+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - The `mqtt` input and output now support sending a last will, configuring a keep alive timeout, and setting retained out output messages.
-- Go API: New stream builder `AddBatchProducerFunc` method.
+- Go API: New stream builder `AddBatchProducerFunc` and `AddBatchConsumerFunc` methods.
 
 ### Fixed